feat: 프로필 조회 기능 추가 (#178)
* feat: 멤버 조회 기능 추가

* feat: 내 정보 조회 기능 추가

* refactor: @Query 사용한 메서드 파라미터에 Param 어노테이션 추가

* feat: 종료된 스터디 조회 기능 추가

* style: 개행 수정

* fix: successRate의 반환타입을 double로 통일

+++ backend/src/main/java/com/yigongil/backend/response/StudyMemberResponse.java
@@ -4,7 +4,7 @@ public record StudyMemberResponse(
         Long id,
         Integer tier,
         String nickname,
-        Integer successRate,
+        Double successRate,
         String profileImage
 ) {
 }
